pilgrims/requirements.txt
Django==1.7.1
psycopg2==2.5.4
django-compressor==1.4
django-debug-toolbar==1.2.2
gunicorn==19.1.1
Fabric==1.10.0
django-localflavor
